requirements.txt
# Core Solana
solana==0.30.2
solders==0.18.1
anchorpy==0.18.0

# Web3 & RPC
httpx==0.25.2
websockets==12.0
aiohttp==3.9.1

# Data & Config
pydantic==2.5.3
pydantic-settings==2.1.0
pyyaml==6.0.1
python-dotenv==1.0.0

# Storage
redis==5.0.1
hiredis==2.3.2

# Dashboard
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.4

# Monitoring & Logging
structlog==23.2.0
python-json-logger==2.0.7

# Security
cryptography==41.0.7
pyage-encryption==1.1.1

# Utilities
click==8.1.7
tenacity==8.2.3
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0

# Development
black==23.12.1
ruff==0.1.9
mypy==1.7.1
